requirements.txt
Django==4.2.2
djangorestframework==3.14.0
psycopg2==2.9.6
Pillow==9.0.0
django-filter==23.2
djoser ==2.2.0
celery[redis]==5.3.0
social-auth-app-django==5.3.0
djangorestframework_simplejwt==5.3.0
django-cors-headers==4.1.0